refactor: refactor component imports and restructure layout for improved organization
