+++ src/app/layout.tsx
@@ -2,7 +2,7 @@ import type { Metadata } from "next";
 import { Geist, Geist_Mono } from "next/font/google";
 import "./globals.css";
 import { Providers } from "./providers";
-import Header from "@/components/Layout/Header";
+import { Header } from "@/components";
 
 const geistSans = Geist({
   variable: "--font-geist-sans",
