+++ src/components/Repository/index.ts
@@ -1,0 +1,1 @@
+export * from "./RepositoryCard"
