+++ src/components/Sections/index.ts
@@ -1,0 +1,2 @@
+export * from "./HomeHeadingSection";
+export * from "./SearchSection";
