+++ src/components/index.ts
@@ -1,0 +1,6 @@
+export * from "./icons";
+export * from "./Frames";
+export * from "./Sections";
+export * from "./Layout";
+export * from "./Profile";
+export * from "./Repository";
